flask-twitter-api/requirements.txt
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
defusedxml==0.7.1
docopt==0.6.2
dominate==2.6.0
Flask==2.0.3
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-HTTPAuth==4.5.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
multidict==6.0.2
pipreqs==0.4.11
Pygments==2.11.2
PyJWT==2.3.0
PySocks==1.7.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.32
urllib3==1.26.9
visitor==0.1.3
Werkzeug==2.0.3
yarg==0.1.9
